almsot got pattenr into meshfill
