+++ Packages/vcs/Lib/meshfill.py
@@ -268,6 +268,7 @@ Class:	Gfm                       	# Meshfill
         'fillareacolors',
         'fillareastyle',
         'fillareaindices',
+        'fillareaopacity',
         'ext_1',
         'ext_2',
         'missing',
@@ -296,6 +297,7 @@ Class:	Gfm                       	# Meshfill
         '_fillareacolors',
         '_fillareastyle',
         '_fillareaindices',
+        '_fillareaopacity',
         '_ext_1',
         '_ext_2',
         '_missing',
@@ -387,6 +389,8 @@ Class:	Gfm                       	# Meshfill
             value)
         self._fillareastyle = value
     fillareastyle = property(_getfillareastyle, _setfillareastyle)
+
+    fillareaopacity = VCS_validation_functions.fillareaopacity
 
     ext_1 = VCS_validation_functions.ext_1
     ext_2 = VCS_validation_functions.ext_2
@@ -609,6 +613,7 @@ Class:	Gfm                       	# Meshfill
             self._fillareastyle = 'solid'
             self._fillareaindices = None
             self._fillareacolors = [1, ]
+            self._fillareaopacity = []
             self._levels = ([1.0000000200408773e+20, 1.0000000200408773e+20],)
             self._legend = None
             self._mesh = 0
@@ -628,7 +633,7 @@ Class:	Gfm                       	# Meshfill
                         'xmtics1', 'xmtics2', 'yticlabels1', 'yticlabels2', 'ymtics1', 'ymtics2',
                         'datawc_y1', 'datawc_y2', 'datawc_x1',
                         'datawc_x2', 'xaxisconvert', 'yaxisconvert', 'missing', 'levels', 'ext_1', 'ext_2',
-                        'fillareastyle', 'fillareaindices', 'fillareacolors', 'legend',
+                        'fillareastyle', 'fillareaindices', 'fillareacolors', 'fillareaopacity', 'legend',
                         'datawc_timeunits', 'datawc_calendar']:
                 setattr(self, "_" + att, getattr(src, "_" + att))
         vcs.elements["meshfill"][Gfm_name] = self
