+++ Packages/vcs/Lib/vcsvtk/boxfillpipeline.py
@@ -1,7 +1,6 @@
 from .pipeline2d import Pipeline2D
 from .. import vcs2vtk
 
-import fillareautils
 import numpy
 import vcs
 import vtk
@@ -307,75 +306,14 @@ class BoxfillPipeline(Pipeline2D):
     def _plotInternalCustomBoxfill(self):
         """Implements the logic to render a custom boxfill."""
         self._mappers = []
-        tmpLevels = []
-        tmpColors = []
-        tmpIndices = []
-        tmpOpacities = []
 
-        indices = self._gm.fillareaindices
-        opacities = self._gm.fillareaopacity
+        prepedContours = self._prepContours()
+        tmpLevels = prepedContours["tmpLevels"]
+        tmpIndices = prepedContours["tmpIndices"]
+        tmpColors = prepedContours["tmpColors"]
+        tmpOpacities = prepedContours["tmpOpacities"]
+
         style = self._gm.fillareastyle
-
-        if indices is None:
-            indices = [1]
-        while len(indices) < len(self._contourColors):
-            indices.append(indices[-1])
-        if len(self._contourLevels) > len(self._contourColors):
-            raise RuntimeError(
-                "You asked for %i levels but provided only %i colors\n"
-                "Graphic Method: %s of type %s\nLevels: %s"
-                % (len(self._contourLevels), len(self._contourColors),
-                   self._gm.name, self._gm.g_name,
-                   repr(self._contourLevels)))
-        elif len(self._contourLevels) < len(self._contourColors) - 1:
-            warnings.warn(
-                "You asked for %i lgridevels but provided %i colors, "
-                "extra ones will be ignored\nGraphic Method: %s of type %s"
-                % (len(self._contourLevels), len(self._contourColors),
-                   self._gm.name, self._gm.g_name))
-        if len(opacities) < len(self._contourColors):
-            # fill up the opacity values
-            if style == 'pattern':
-                opacities += [None] * (len(self._contourColors) - len(opacities))
-            else:
-                opacities += [None] * (len(self._contourColors) - len(opacities))
-
-        # The following loop attempts to group isosurfaces based on their
-        # attributes. Isosurfaces grouped if and only if all properties match.
-        for i, l in enumerate(self._contourLevels):
-            if i == 0:
-                C = [self._contourColors[i]]
-                if numpy.allclose(self._contourLevels[0][0], -1.e20):
-                    # ok it's an extension arrow
-                    L = [self._scalarRange[0] - 1., self._contourLevels[0][1]]
-                else:
-                    L = list(self._contourLevels[i])
-                I = indices[i]
-                O = opacities[i]
-            else:
-                if l[0] == L[-1] and I == indices[i] and\
-                        ((style == 'solid') or
-                            (C[-1] == self._contourColors[i] and O == opacities[i])):
-                    # Ok same type lets keep going
-                    if numpy.allclose(l[1], 1.e20):
-                        L.append(self._scalarRange[1] + 1.)
-                    else:
-                        L.append(l[1])
-                    C.append(self._contourColors[i])
-                else:  # ok we need new contouring
-                    tmpLevels.append(L)
-                    tmpColors.append(C)
-                    tmpIndices.append(I)
-                    tmpOpacities.append(O)
-                    C = [self._contourColors[i]]
-#                    L = self._contourLevels[i]
-                    L = [L[-1], l[1]]
-                    I = indices[i]
-                    O = opacities[i]
-        tmpLevels.append(L)
-        tmpColors.append(C)
-        tmpIndices.append(I)
-        tmpOpacities.append(O)
 
         luts = []
         geos = []
@@ -422,19 +360,7 @@ class BoxfillPipeline(Pipeline2D):
                     self._mappers.append(mapper)
 
                 #  Since pattern creation requires a single color, assuming the first
-                c = [val * 255 / 100.0 for val in _colorMap.index[tmpColors[i][0]]]
-                tmpOpacity = tmpOpacities[i]
-                if tmpOpacity is None:
-                    tmpOpacity = c[-1]
-                else:
-                    tmpOpacity = tmpOpacity * 255 / 100.
-                act = fillareautils.make_patterned_polydata(geoFilter2.GetOutput(),
-                                                            fillareastyle=style,
-                                                            fillareaindex=tmpIndices[i],
-                                                            fillareacolors=c,
-                                                            fillareaopacity=tmpOpacity)
-                if act is not None:
-                    self._patternActors.append(act)
+                self._patternCreation(geoFilter2,_colorMap.index[tmpColors[i][0]],style,tmpIndices[i],tmpOpacities[i])
 
         self._resultDict["vtk_backend_luts"] = luts
         if len(geos) > 0:
